pyproject.toml
[tool.poetry]
name = "zombiecoder-mcp"
version = "0.1.0"
description = "ZombieCoder MCP System"
authors = ["ZombieCoder Team"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.0.1"
websockets = "^10.0"
psutil = "^5.8.0"
pyttsx3 = "^2.90"
gTTS = "^2.2.3"
PyYAML = "^6.0"
requests = "^2.26.0"
python-dotenv = "^0.19.0"
aiohttp = "^3.8.1"
watchdog = "^2.1.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
